feat: Updating Resume Format - Trying a new resume format/layout - Upgrading all deps to latest - Changing comma to semicolon in LinkedIn context menu helper
