+++ api/package.json
@@ -21,36 +21,36 @@
     "explode": "npx rimraf node_modules package-lock.json && npm i"
   },
   "dependencies": {
-    "@nestjs/common": "^10.4.3",
+    "@nestjs/common": "^10.4.5",
     "@nestjs/config": "^3.2.3",
-    "@nestjs/core": "^10.4.3",
+    "@nestjs/core": "^10.4.5",
     "@nestjs/jwt": "^10.2.0",
-    "@nestjs/platform-express": "^10.4.3",
+    "@nestjs/platform-express": "^10.4.5",
     "@nestjs/sequelize": "^10.0.1",
     "@nestjs/swagger": "^7.4.2",
     "just-diff": "^6.0.2",
     "lodash": "^4.17.21",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
-    "sequelize": "^6.37.3",
+    "sequelize": "^6.37.4",
     "sequelize-typescript": "^2.1.6",
     "sqlite3": "^5.1.7",
     "uuidv4": "^6.2.13"
   },
   "devDependencies": {
-    "@faker-js/faker": "^9.0.1",
+    "@faker-js/faker": "^9.0.3",
     "@nestjs/cli": "^10.4.5",
-    "@nestjs/schematics": "^10.1.4",
-    "@nestjs/testing": "^10.4.3",
-    "@openapitools/openapi-generator-cli": "^2.13.7",
-    "@types/express": "^4.17.21",
+    "@nestjs/schematics": "^10.2.2",
+    "@nestjs/testing": "^10.4.5",
+    "@openapitools/openapi-generator-cli": "^2.14.0",
+    "@types/express": "^5.0.0",
     "@types/jest": "^29.5.13",
-    "@types/lodash": "^4.17.7",
-    "@types/node": "^22.5.5",
+    "@types/lodash": "^4.17.11",
+    "@types/node": "^22.7.6",
     "@types/sequelize": "^4.28.20",
     "@types/supertest": "^6.0.2",
-    "@typescript-eslint/eslint-plugin": "^8.6.0",
-    "@typescript-eslint/parser": "^8.6.0",
+    "@typescript-eslint/eslint-plugin": "^8.10.0",
+    "@typescript-eslint/parser": "^8.10.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "jest": "^29.7.0",
@@ -61,7 +61,7 @@
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.3"
   },
   "jest": {
     "moduleFileExtensions": [
