+++ linkedIn/details2.js
@@ -1,1 +1,1 @@
-navigator.clipboard.writeText(`${document.getElementsByClassName('job-details-jobs-unified-top-card__company-name')[0].getElementsByClassName('app-aware-link ')[0].innerText},${document.getElementsByClassName('t-24 t-bold inline')[0].innerText},${new Intl.DateTimeFormat('en-US').format(new Date())},${location.href.split('?')[0]}`);
+navigator.clipboard.writeText(`${document.getElementsByClassName('job-details-jobs-unified-top-card__company-name')[0].getElementsByClassName('app-aware-link ')[0].innerText};${document.getElementsByClassName('t-24 t-bold inline')[0].innerText};${new Intl.DateTimeFormat('en-US').format(new Date())};${location.href.split('?')[0]}`);
