+++ resume/package.json
@@ -9,19 +9,23 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "vue": "^3.5.6"
+    "vue": "^3.5.12"
   },
   "devDependencies": {
     "@incutonez/eslint-plugin": "^1.1.4",
-    "@stylistic/eslint-plugin-ts": "^2.8.0",
-    "@typescript-eslint/eslint-plugin": "^8.6.0",
-    "@typescript-eslint/parser": "^8.6.0",
+    "@stylistic/eslint-plugin-ts": "^2.9.0",
+    "@typescript-eslint/eslint-plugin": "^8.10.0",
+    "@typescript-eslint/parser": "^8.10.0",
     "@vitejs/plugin-vue": "^5.1.4",
     "autoprefixer": "^10.4.20",
+    "esbuild": "^0.21.5",
+    "eslint-plugin-tailwindcss": "^3.17.5",
+    "eslint-plugin-vue": "^9.29.0",
     "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.12",
-    "typescript": "^5.6.2",
-    "vite": "^5.4.6",
+    "rollup": "^4.24.0",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.9",
     "vite-svg-loader": "^5.1.0",
     "vue-tsc": "^2.1.6"
   }
