--- resume/src/components/SectionEducation.vue
@@ -1,37 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-	<article>
-		<h1 class="section-title">
-			Education
-		</h1>
-		<section class="mt-1">
-			<div class="font-semibold">
-				Bachelor of Computer Science
-			</div>
-			<div>
-				Penn State
-			</div>
-			<div>
-				Aug 2006 - May 2010
-			</div>
-		</section>
-		<section class="mt-4">
-			<div class="font-semibold">
-				4 Postgraduate Classes
-			</div>
-			<div>
-				UMBC
-			</div>
-			<div>
-				Aug 2010 - May 2012
-			</div>
-		</section>
-	</article>
-</template>
-
-<style scoped lang="scss">
-
-</style>
