+++ resume/src/components/SectionJobDescription.vue
@@ -22,20 +22,20 @@ defineProps<ISectionJobDescription>();
 			<span class="font-semibold">{{ company }}</span>
 			<span class="mx-1">|</span>
 			<span>{{ location }}</span>
-			<span class="w-full font-semibold md:ml-auto md:w-auto">{{ dateStart }} - {{ dateEnd }}</span>
+			<span class="w-full font-semibold text-gray-700 md:ml-auto md:w-auto">{{ dateStart }} - {{ dateEnd }}</span>
 		</section>
 		<section class="mb-0.5 flex flex-col text-sm md:flex-row">
 			<span>
 				<span class="font-semibold text-blue-resume">Project{{ multipleProjects ? 's' : '' }}</span>
 				<span>: {{ project }}</span>
 			</span>
-			<span class="italic md:ml-auto">{{ industry }}</span>
+			<span class="italic text-gray-700 md:ml-auto">{{ industry }}</span>
 		</section>
 		<template v-if="!!tasks">
 			<ul
 				v-for="task in tasks"
 				:key="task"
-				class="ml-4 list-disc text-xs"
+				class="ml-4 list-disc text-sm"
 			>
 				<li>{{ task }}</li>
 			</ul>
