--- resume/src/components/SectionSummary.vue
@@ -1,18 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-	<article>
-		<h1 class="section-title">
-			Summary
-		</h1>
-		<p class="text-sm">
-			Meticulous Senior Web Engineer with a big personality looking for an even bigger project to wrangle.  Loves long walks on the beach with Vue but has a curiosity for learning React or Angular.  Mentors engineers of all ages, ranks, and walks of life.  Jira champion, enthusiastic debugger, defensive programmer, effective communicator, and a loud advocator for best practices.
-		</p>
-	</article>
-</template>
-
-<style scoped lang="scss">
-
-</style>
