+++ resume/src/style.css
@@ -14,6 +14,6 @@
 
 @layer components {
     .section-title {
-        @apply text-lg font-semibold;
+        @apply text-lg font-semibold border-b-2 border-black;
     }
 }
