+++ spec/package.json
@@ -11,6 +11,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.3"
   }
 }
