+++ ui/package.json
@@ -10,18 +10,17 @@
     "game:copyEnums": "npx tsx copyEnums.ts"
   },
   "dependencies": {
-    "@faker-js/faker": "^9.0.1",
-    "@material-symbols/svg-400": "^0.23.0",
-    "@sinclair/typebox": "^0.33.12",
-    "@tiptap/pm": "2.6.6",
-    "@tiptap/starter-kit": "2.6.6",
-    "@tiptap/vue-3": "2.6.6",
+    "@faker-js/faker": "^9.0.3",
+    "@material-symbols/svg-400": "^0.25.1",
+    "@sinclair/typebox": "^0.33.17",
+    "@tiptap/pm": "2.8.0",
+    "@tiptap/starter-kit": "2.8.0",
+    "@tiptap/vue-3": "2.8.0",
     "@types/pluralize": "^0.0.33",
     "@vue/language-core": "^2.1.6",
     "@vueuse/core": "^11.1.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
-    "esbuild": "^0.23.1",
     "glob": "^11.0.0",
     "just-clone": "^6.2.0",
     "just-is-empty": "^3.4.1",
@@ -32,32 +31,33 @@
     "papaparse": "^5.4.1",
     "path-browserify": "^1.0.1",
     "pluralize": "^8.0.0",
-    "primevue": "^4.0.7",
-    "rollup": "^4.22.0",
-    "vue": "^3.5.6",
+    "primevue": "^4.1.1",
+    "vue": "^3.5.12",
     "vue-router": "^4.4.5",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@incutonez/eslint-plugin": "^1.1.4",
-    "@stylistic/eslint-plugin-ts": "^2.8.0",
+    "@stylistic/eslint-plugin-ts": "^2.9.0",
     "@types/glob": "^8.1.0",
     "@types/lodash-es": "^4.17.12",
     "@types/mime-types": "^2.1.4",
     "@types/papaparse": "^5.3.14",
-    "@typescript-eslint/eslint-plugin": "^8.6.0",
-    "@typescript-eslint/parser": "^8.6.0",
+    "@typescript-eslint/eslint-plugin": "^8.10.0",
+    "@typescript-eslint/parser": "^8.10.0",
     "@vitejs/plugin-vue": "^5.1.4",
     "autoprefixer": "^10.4.20",
-    "eslint-plugin-tailwindcss": "^3.17.4",
-    "eslint-plugin-vue": "^9.28.0",
-    "globals": "^15.9.0",
+    "esbuild": "^0.24.0",
+    "eslint-plugin-tailwindcss": "^3.17.5",
+    "eslint-plugin-vue": "^9.29.0",
+    "globals": "^15.11.0",
     "postcss": "^8.4.47",
     "reflect-metadata": "^0.2.2",
-    "sass": "^1.79.1",
-    "tailwindcss": "^3.4.12",
-    "typescript": "^5.6.2",
-    "vite": "^5.4.6",
+    "rollup": "^4.24.0",
+    "sass": "^1.80.2",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.9",
     "vite-svg-loader": "^5.1.0",
     "vue-tsc": "^2.1.6"
   }
